Split commands/get.rs and commands/helpers.rs into submodules
Refs #107.

commands/get.rs (1770 lines) → commands/get/ (5 files):
  - mod.rs: PadFilter, run() orchestrator
  - status_filter.rs: PadStatusFilter, filter_tree, filter_pad_recursive,
    filter_children_under_deleted, matches_status
  - attr_filter.rs: apply_attr_filters, filter_pad_by_attrs
  - selector_filter.rs: filter_by_selectors, linearize_*, find_by_path
  - search.rs: apply_search, highlight_matches, extract_context

commands/helpers.rs (1441 lines) → commands/helpers/ (6 files):
  - mod.rs: fmt_path, re-exports
  - indexing.rs: indexed_pads, bucket_for_index
  - selector_resolve.rs: resolve_selectors, linearize_*, check_protection
    (factored out duplicated delete-protection check from the 5 selector arms)
  - pad_fetching.rs: pads_by_selectors
  - nesting.rs: collect_nested_pads, NestedPad, flatten_tree
  - tree_search.rs: get_descendant_ids, find_pad_by_uuid, find_node_by_id,
    collect_subtree_ids

All public items re-exported from mod.rs; no changes to external callers.
Tests travel with their subject matter.
